+++ include/GRMustacheInvocation.h
@@ -28,6 +28,6 @@
     id _returnValue;
     id _token;
 }
-@property (nonatomic, readonly) NSString *key AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-@property (nonatomic, retain) id returnValue AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
+@property (nonatomic, readonly) NSString *key AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
+@property (nonatomic, retain) id returnValue AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 @end