+++ include/GRMustacheTemplate.h
@@ -33,11 +33,10 @@
 @interface GRMustacheTemplate: NSObject {
 @private
     NSArray *_elems;
-    GRMustacheTemplateOptions _options;
     id<GRMustacheTemplateDelegate> _delegate;
 }
 
-@property (nonatomic, assign) id<GRMustacheTemplateDelegate> delegate AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
+@property (nonatomic, assign) id<GRMustacheTemplateDelegate> delegate AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -52,25 +51,10 @@
  @return A GRMustacheTemplate instance
  @param templateString The template string
  @param outError If there is an error loading or parsing template and partials, upon return contains an NSError object that describes the problem.
- @see templateFromString:options:error:
  @see [GRMustache defaultTemplateOptions]
  @since v1.11
  */
-+ (id)templateFromString:(NSString *)templateString error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-
-/**
- Parses a template string, and returns a compiled template.
- 
- The behavior of the returned template is determined by the _options_ parameter, not by [GRMustache defaultTemplateOptions].
- 
- @return A GRMustacheTemplate instance
- @param templateString The template string
- @param options A mask of options indicating the behavior of the template.
- @param outError If there is an error loading or parsing template and partials, upon return contains an NSError object that describes the problem.
- @see templateFromString:error:
- @since v1.11
- */
-+ (id)templateFromString:(NSString *)templateString options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (id)templateFromString:(NSString *)templateString error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 /**
  Renders a context object from a template string.
@@ -82,8 +66,7 @@
  
  @since v1.0
  */
-+ (NSString *)renderObject:(id)object fromString:(NSString *)templateString error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (NSString *)renderObject:(id)object fromString:(NSString *)templateString options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (NSString *)renderObject:(id)object fromString:(NSString *)templateString error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -101,8 +84,7 @@
  
  @since v1.11
  */
-+ (id)templateFromContentsOfFile:(NSString *)path error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (id)templateFromContentsOfFile:(NSString *)path options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (id)templateFromContentsOfFile:(NSString *)path error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 #if !TARGET_OS_IPHONE || __IPHONE_OS_VERSION_MAX_ALLOWED >= 40000
 
@@ -117,8 +99,7 @@
  
  @since v1.11
  */
-+ (id)templateFromContentsOfURL:(NSURL *)url error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (id)templateFromContentsOfURL:(NSURL *)url options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (id)templateFromContentsOfURL:(NSURL *)url error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 #endif /* if !TARGET_OS_IPHONE || __IPHONE_OS_VERSION_MAX_ALLOWED >= 40000 */
 
@@ -134,8 +115,7 @@
  
  @since v1.4
  */
-+ (NSString *)renderObject:(id)object fromContentsOfFile:(NSString *)path error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (NSString *)renderObject:(id)object fromContentsOfFile:(NSString *)path options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (NSString *)renderObject:(id)object fromContentsOfFile:(NSString *)path error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 
 #if !TARGET_OS_IPHONE || __IPHONE_OS_VERSION_MAX_ALLOWED >= 40000
@@ -152,8 +132,7 @@
  
  @since v1.0
  */
-+ (NSString *)renderObject:(id)object fromContentsOfURL:(NSURL *)url error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (NSString *)renderObject:(id)object fromContentsOfURL:(NSURL *)url options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (NSString *)renderObject:(id)object fromContentsOfURL:(NSURL *)url error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 #endif /* if !TARGET_OS_IPHONE || __IPHONE_OS_VERSION_MAX_ALLOWED >= 40000 */
 
@@ -176,8 +155,7 @@
  
  @since v1.11
  */
-+ (id)templateFromResource:(NSString *)name bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (id)templateFromResource:(NSString *)name bundle:(NSBundle *)bundle options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (id)templateFromResource:(NSString *)name bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 /**
  Parses a bundle resource template, and returns a compiled template.
@@ -194,8 +172,7 @@
  
  @since v1.11
  */
-+ (id)templateFromResource:(NSString *)name withExtension:(NSString *)ext bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (id)templateFromResource:(NSString *)name withExtension:(NSString *)ext bundle:(NSBundle *)bundle options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (id)templateFromResource:(NSString *)name withExtension:(NSString *)ext bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 
 /**
@@ -213,8 +190,7 @@
  
  @since v1.0
  */
-+ (NSString *)renderObject:(id)object fromResource:(NSString *)name bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (NSString *)renderObject:(id)object fromResource:(NSString *)name bundle:(NSBundle *)bundle options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (NSString *)renderObject:(id)object fromResource:(NSString *)name bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 /**
  Renders a context object from a bundle resource template.
@@ -232,8 +208,7 @@
  
  @since v1.0
  */
-+ (NSString *)renderObject:(id)object fromResource:(NSString *)name withExtension:(NSString *)ext bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
-+ (NSString *)renderObject:(id)object fromResource:(NSString *)name withExtension:(NSString *)ext bundle:(NSBundle *)bundle options:(GRMustacheTemplateOptions)options error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
++ (NSString *)renderObject:(id)object fromResource:(NSString *)name withExtension:(NSString *)ext bundle:(NSBundle *)bundle error:(NSError **)outError AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -248,7 +223,7 @@
  
  @since v1.0
  */
-- (NSString *)renderObject:(id)object AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
+- (NSString *)renderObject:(id)object AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 /**
  Renders a template with context objects.
@@ -258,7 +233,7 @@
  
  @since v1.5
  */
-- (NSString *)renderObjects:(id)object, ... __attribute__ ((sentinel)) AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
+- (NSString *)renderObjects:(id)object, ... __attribute__ ((sentinel)) AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 /**
  Renders a template without any context object for interpreting Mustache tags.
@@ -267,6 +242,6 @@
  
  @since v1.0
  */
-- (NSString *)render AVAILABLE_GRMUSTACHE_VERSION_2_0_AND_LATER;
+- (NSString *)render AVAILABLE_GRMUSTACHE_VERSION_3_0_AND_LATER;
 
 @end